Check chrome browser and version
