+++ adyenv6b2ccheckoutaddon/acceleratoraddon/web/src/com/adyen/v6/utils/SameSiteCookieAttributeAppenderUtils.java
@@ -25,6 +25,7 @@ import com.google.common.net.HttpHeaders;
 import de.hybris.platform.servicelayer.config.ConfigurationService;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.configuration.Configuration;
+import org.apache.log4j.Logger;
 
 import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletRequest;
@@ -34,8 +35,12 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class SameSiteCookieAttributeAppenderUtils {
+
+    private static final Logger LOG = Logger.getLogger(SameSiteCookieAttributeAppenderUtils.class);
 
     private ConfigurationService configurationService;
 
@@ -43,7 +48,7 @@ public class SameSiteCookieAttributeAppenderUtils {
     private static final String SAMESITE_COOKIE_HANDLER_ENABLED_PROPERTY = "adyen.samesitecookie.handler.enabled";
     private static final List<String> SAP_VERSIONS_WITH_SAMESITE_FIX = Collections.singletonList("2005");
     private static final List<String> COOKIES_WITH_FORCE_SAME_SITE_NONE = Arrays.asList("JSESSIONID", "acceleratorSecureGUID");
-
+    private static final Pattern CHROME_VERSION = Pattern.compile("Chrom[^ \\/]+\\/(\\d+)[\\.\\d]*");
 
     public void addSameSiteAttribute(HttpServletRequest servletRequest, HttpServletResponse servletResponse) {
         // Do not modify cookies for SAP versions which already have SameSite cookies handler available
@@ -71,9 +76,7 @@ public class SameSiteCookieAttributeAppenderUtils {
                 .filter(cookie -> cookie.startsWith(sameSiteCookie) && !cookie.contains("SameSite"))
                 .findAny();
 
-        if (sessionCookieWithoutSameSite.isPresent()
-                //&& SameSiteCookieUtils.shouldSendSameSiteNone(userAgent)
-        ) {
+        if (sessionCookieWithoutSameSite.isPresent() && shouldSendSameSiteNone(userAgent)) {
             // Replace all set cookie headers with 1 new session + sameSite header
             servletResponse.setHeader(HttpHeaders.SET_COOKIE, sessionCookieWithoutSameSite.get() + ";Secure ;SameSite=None");
 
@@ -95,6 +98,27 @@ public class SameSiteCookieAttributeAppenderUtils {
             isSameSiteCookieHandlingEnabled = configuration.getBoolean(SAMESITE_COOKIE_HANDLER_ENABLED_PROPERTY);
         }
         return isSameSiteCookieHandlingEnabled;
+    }
+
+    public static boolean shouldSendSameSiteNone(String useragent) {
+        return isChromiumBased(useragent) && isChromiumVersionAtLeast(80, useragent);
+    }
+
+    private static boolean isChromiumBased(String useragent) {
+        return useragent.contains("Chrome") || useragent.contains("Chromium");
+    }
+
+    private static boolean isChromiumVersionAtLeast(int major, String useragent) {
+        Matcher matcher = CHROME_VERSION.matcher(useragent);
+        if (matcher.find()) {
+            try {
+                String chromeVersion = matcher.group(1);
+                return Integer.parseInt(chromeVersion) >= major;
+            } catch (NumberFormatException ignored) {
+                LOG.debug("Could not parse Chrome browser version, SameSite cookie handling will be skipped");
+            }
+        }
+        return false;
     }
 
     public ConfigurationService getConfigurationService() {
